Fix bugs: 56318  function literals with names don't work right 57045  negative integers as object properties: weird behavior 58479  functions defined within conditional phrases are always crea

+++ js/rhino/org/mozilla/javascript/InterpreterData.java
@@ -80,7 +80,7 @@ class InterpreterData {
     
     public boolean removeBreakpoint(int line) {
         int offset = getOffset(line);
-        if (offset != -1 && itsICode[offset] == TokenStream.BREAKPOINT)
+        if (offset != -1 && itsICode[offset] == (byte) TokenStream.BREAKPOINT)
         {
             itsICode[offset] = (byte) TokenStream.LINE;
             return true;
